+++ Sources/VCore/Extensions/Foundation/Collection/Collection.NonEmpty.swift
@@ -7,7 +7,7 @@
 
 import Foundation
 
-// MARK: - Collection Non Empty
+// MARK: - Collection Non-Empty
 extension Collection {
     /// Returns non-empty `Collection`, or `nil`.
     ///