+++ Sources/VCore/Extensions/Foundation/Data/Data.NonEmpty.swift
@@ -7,7 +7,7 @@
 
 import Foundation
 
-// MARK: - Data Non Empty
+// MARK: - Data Non-Empty
 extension Data {
     /// Returns non-zero sized `Data`, or `nil`.
     ///