+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Extensions
 
 - `Array.removeIfPresent(_:)` and `Array.removeIfPresent(contentsOf:)` methods are added, that remove elements if they are present
 - `Collection.nonEmpty` property is added, that returns non-empty `Collection`, or `nil`.
+- `Numeric.nonZero` property is added, that return non-zero sized `Numeric`, or `nil`.
 - `withContinuousObservationTracking(of:on:execute:)` method is added, that tracks access to properties in a `class`-type continuously
 - `View.inlineNavigationTitle(_:)` method is added with `LocalizedStringKey`